+++ packages/apps/navigator/src/components/RoutesStyle.tsx
@@ -1,7 +1,55 @@
+import { Preconditions } from '@truckermudgeon/base/precon';
 import { routingModes } from '@truckermudgeon/map/routing';
-import type { DataDrivenPropertyValueSpecification } from 'maplibre-gl';
+import type {
+  DataDrivenPropertyValueSpecification,
+  ExpressionSpecification,
+} from 'maplibre-gl';
 import { Layer, Source } from 'react-map-gl/maplibre';
-import { lineGradientExpression } from './SlippyMap';
+
+const lineColors = {
+  case: {
+    before: 'hsl(204,0%,60%)',
+    after: 'hsl(204,100%,40%)',
+  },
+  line: {
+    before: 'hsl(204,0%,80%)',
+    after: 'hsl(204,100%,50%)',
+  },
+  animatedPrimaryCase: {
+    before: 'hsl(204,100%,40%)',
+    after: 'rgba(0, 0, 0, 0)',
+  },
+  animatedPrimaryLine: {
+    before: 'hsl(204,100%,50%)',
+    after: 'rgba(0, 0, 0, 0)',
+  },
+  animatedSecondaryCase: {
+    before: 'hsl(204,80%,70%)',
+    after: 'rgba(0, 0, 0, 0)',
+  },
+  animatedSecondaryLine: {
+    before: 'hsl(204,80%,80%)',
+    after: 'rgba(0, 0, 0, 0)',
+  },
+};
+
+export const lineGradientExpression = ({
+  lineType,
+  progress,
+}: {
+  lineType: keyof typeof lineColors;
+  progress: number;
+}) => {
+  Preconditions.checkArgument(0 <= progress && progress <= 1);
+  const { before, after } = lineColors[lineType];
+  return [
+    'step',
+    ['line-progress'],
+    before,
+    progress,
+    after,
+  ] satisfies ExpressionSpecification;
+};
 
 export const RoutesStyle = () => {
   console.log('render routes layers');
@@ -70,7 +118,6 @@ export const RoutesStyle = () => {
         />
         <Layer
           id={'activeRouteStepLayer'}
-          source={'activeRouteStep'}
           type={'line'}
           paint={{
             'line-width': routeLineWidth,
@@ -84,7 +131,7 @@ export const RoutesStyle = () => {
       </Source>
       <Layer
         id={'activeRouteIconsLayer'}
-        source={'activeRouteStep'}
+        source={'activeRoute'}
         type={'symbol'}
         layout={{
           'icon-image': '{sprite}',
