+++ packages/apps/navigator/src/components/SlippyMap.tsx
@@ -1,17 +1,13 @@
 import { assertExists } from '@truckermudgeon/base/assert';
 import type { Extent } from '@truckermudgeon/base/geom';
 import { center as calculateCenter } from '@truckermudgeon/base/geom';
-import { Preconditions } from '@truckermudgeon/base/precon';
 import {
   BaseMapStyle,
   defaultMapStyle,
   GameMapStyle,
   SceneryTownSource,
 } from '@truckermudgeon/ui';
-import type {
-  ExpressionSpecification,
-  Marker as MapLibreGLMarker,
-} from 'maplibre-gl';
+import type { Marker as MapLibreGLMarker } from 'maplibre-gl';
 import type { ForwardRefExoticComponent, ReactElement } from 'react';
 import { useRef } from 'react';
 import type { MapRef } from 'react-map-gl/maplibre';
@@ -32,51 +28,6 @@ const extents = {
     [-10.025698, 34.897275].map(n => Math.floor(n)),
     [33.284941, 71.573102].map(n => Math.ceil(n)),
   ].flat() as Extent,
-};
-
-const lineColors = {
-  case: {
-    before: 'hsl(204,0%,60%)',
-    after: 'hsl(204,100%,40%)',
-  },
-  line: {
-    before: 'hsl(204,0%,80%)',
-    after: 'hsl(204,100%,50%)',
-  },
-  animatedPrimaryCase: {
-    before: 'hsl(204,100%,40%)',
-    after: 'rgba(0, 0, 0, 0)',
-  },
-  animatedPrimaryLine: {
-    before: 'hsl(204,100%,50%)',
-    after: 'rgba(0, 0, 0, 0)',
-  },
-  animatedSecondaryCase: {
-    before: 'hsl(204,80%,70%)',
-    after: 'rgba(0, 0, 0, 0)',
-  },
-  animatedSecondaryLine: {
-    before: 'hsl(204,80%,80%)',
-    after: 'rgba(0, 0, 0, 0)',
-  },
-};
-
-export const lineGradientExpression = ({
-  lineType,
-  progress,
-}: {
-  lineType: keyof typeof lineColors;
-  progress: number;
-}) => {
-  Preconditions.checkArgument(0 <= progress && progress <= 1);
-  const { before, after } = lineColors[lineType];
-  return [
-    'step',
-    ['line-progress'],
-    before,
-    progress,
-    after,
-  ] satisfies ExpressionSpecification;
 };
 
 export const SlippyMap = (props: {
